Split sap system database aggregates (#2445)
* Add new database aggregate code

* Remove database logic from sap system aggregate

* Redirect router commands to database aggregate

* Adapt deregistration process manager

* Adapt enrichment protocol

* Adapt sap system discovery policy

* Adapt remaining code alias and tests

* Adapt related e2e tests

* Update the aggregates docstring

* Leftover dosctring and pattern matching entries
