+++ lib/trento/databases/commands/deregister_database_instance.ex
@@ -1,4 +1,4 @@
-defmodule Trento.SapSystems.Commands.DeregisterDatabaseInstance do
+defmodule Trento.Databases.Commands.DeregisterDatabaseInstance do
   @moduledoc """
   Deregister (decommission) a database instance from the monitoring system.
   """
@@ -10,7 +10,7 @@ defmodule Trento.SapSystems.Commands.DeregisterDatabaseInstance do
   defcommand do
     field :instance_number, :string
     field :host_id, Ecto.UUID
-    field :sap_system_id, Ecto.UUID
+    field :database_id, Ecto.UUID
     field :deregistered_at, :utc_datetime_usec
   end
 end
