+++ lib/trento/infrastructure/commanded/middleware/enrich_register_application_instance.ex
@@ -18,7 +18,8 @@ defimpl Trento.Infrastructure.Commanded.Middleware.Enrichable,
 
     case Repo.one(query) do
       %DatabaseInstanceReadModel{sap_system_id: sap_system_id} ->
-        {:ok, %RegisterApplicationInstance{command | sap_system_id: sap_system_id}}
+        {:ok,
+         %RegisterApplicationInstance{command | sap_system_id: UUID.uuid5(sap_system_id, tenant)}}
 
       nil ->
         {:error, :database_not_registered}
