+++ test/trento/infrastructure/commanded/middleware/enrich_register_application_instance_test.exs
@@ -27,7 +27,9 @@ defmodule Trento.Infrastructure.Commanded.Middleware.EnrichRegisterApplicationIn
         health: :passing
       )
 
-    assert {:ok, %RegisterApplicationInstance{sap_system_id: ^sap_system_id}} =
+    expected_sap_system_id = UUID.uuid5(sap_system_id, tenant)
+
+    assert {:ok, %RegisterApplicationInstance{sap_system_id: ^expected_sap_system_id}} =
              Enrichable.enrich(command, %{})
   end
