+++ test/trento/sap_systems/sap_system_test.exs
@@ -7,11 +7,8 @@ defmodule Trento.SapSystems.SapSystemTest do
 
   alias Trento.SapSystems.Commands.{
     DeregisterApplicationInstance,
-    DeregisterDatabaseInstance,
     MarkApplicationInstanceAbsent,
-    MarkDatabaseInstanceAbsent,
     RegisterApplicationInstance,
-    RegisterDatabaseInstance,
     RollUpSapSystem
   }
 
@@ -22,16 +19,6 @@ defmodule Trento.SapSystems.SapSystemTest do
     ApplicationInstanceMarkedPresent,
     ApplicationInstanceMoved,
     ApplicationInstanceRegistered,
-    DatabaseDeregistered,
-    DatabaseHealthChanged,
-    DatabaseInstanceDeregistered,
-    DatabaseInstanceHealthChanged,
-    DatabaseInstanceMarkedAbsent,
-    DatabaseInstanceMarkedPresent,
-    DatabaseInstanceRegistered,
-    DatabaseInstanceSystemReplicationChanged,
-    DatabaseRegistered,
-    DatabaseRestored,
     SapSystemDeregistered,
     SapSystemHealthChanged,
     SapSystemRegistered,
@@ -43,311 +30,11 @@ defmodule Trento.SapSystems.SapSystemTest do
   }
 
   alias Trento.SapSystems.{
-    Application,
-    Database,
     Instance,
     SapSystem
   }
 
   describe "SAP System registration" do
-    test "should fail when a sap system does not exists and the database instance has Secondary role" do
-      command =
-        build(:register_database_instance_command,
-          system_replication: "Secondary"
-        )
-
-      assert_error(
-        command,
-        {:error, :sap_system_not_registered}
-      )
-    end
-
-    test "should create an incomplete SAP system aggregate and register a database instance when the system replication is disabled" do
-      sap_system_id = Faker.UUID.v4()
-      sid = Faker.StarWars.planet()
-      tenant = Faker.Beer.style()
-      instance_number = "00"
-      instance_hostname = Faker.Airports.iata()
-      features = Faker.Pokemon.name()
-      http_port = 80
-      https_port = 443
-      start_priority = "0.9"
-      host_id = Faker.UUID.v4()
-
-      assert_events_and_state(
-        [],
-        RegisterDatabaseInstance.new!(%{
-          sap_system_id: sap_system_id,
-          sid: sid,
-          tenant: tenant,
-          instance_number: instance_number,
-          instance_hostname: instance_hostname,
-          features: features,
-          http_port: http_port,
-          https_port: https_port,
-          start_priority: start_priority,
-          host_id: host_id,
-          system_replication: nil,
-          health: :passing
-        }),
-        [
-          %DatabaseRegistered{
-            sap_system_id: sap_system_id,
-            sid: sid,
-            health: :passing
-          },
-          %DatabaseInstanceRegistered{
-            sap_system_id: sap_system_id,
-            sid: sid,
-            tenant: tenant,
-            instance_number: instance_number,
-            instance_hostname: instance_hostname,
-            features: features,
-            http_port: http_port,
-            https_port: https_port,
-            start_priority: start_priority,
-            host_id: host_id,
-            system_replication: nil,
-            system_replication_status: nil,
-            health: :passing
-          }
-        ],
-        %SapSystem{
-          sap_system_id: sap_system_id,
-          # The SAP System aggregate is not complete yet.
-          # The sid will be set when the first application instance is registered.
-          sid: nil,
-          database: %Database{
-            sid: sid,
-            health: :passing,
-            instances: [
-              %Instance{
-                sid: sid,
-                system_replication: nil,
-                system_replication_status: nil,
-                instance_number: instance_number,
-                features: features,
-                host_id: host_id,
-                health: :passing,
-                absent_at: nil
-              }
-            ]
-          }
-        }
-      )
-    end
-
-    test "should create an incomplete SAP system aggregate and register a database instance when the system replication is enabled and the database role is primary" do
-      sap_system_id = Faker.UUID.v4()
-      sid = Faker.StarWars.planet()
-      tenant = Faker.Beer.style()
-      instance_number = "00"
-      instance_hostname = Faker.Airports.iata()
-      features = Faker.Pokemon.name()
-      http_port = 80
-      https_port = 443
-      start_priority = "0.9"
-      host_id = Faker.UUID.v4()
-
-      assert_events_and_state(
-        [],
-        RegisterDatabaseInstance.new!(%{
-          sap_system_id: sap_system_id,
-          sid: sid,
-          tenant: tenant,
-          instance_number: instance_number,
-          instance_hostname: instance_hostname,
-          features: features,
-          http_port: http_port,
-          https_port: https_port,
-          start_priority: start_priority,
-          host_id: host_id,
-          system_replication: "Primary",
-          system_replication_status: "ACTIVE",
-          health: :passing
-        }),
-        [
-          %DatabaseRegistered{
-            sap_system_id: sap_system_id,
-            sid: sid,
-            health: :passing
-          },
-          %DatabaseInstanceRegistered{
-            sap_system_id: sap_system_id,
-            sid: sid,
-            tenant: tenant,
-            instance_number: instance_number,
-            instance_hostname: instance_hostname,
-            features: features,
-            http_port: http_port,
-            https_port: https_port,
-            start_priority: start_priority,
-            host_id: host_id,
-            system_replication: "Primary",
-            system_replication_status: "ACTIVE",
-            health: :passing
-          }
-        ],
-        %SapSystem{
-          sap_system_id: sap_system_id,
-          # The SAP System aggregate is not complete yet.
-          # The sid will be set when the first application instance is registered.
-          sid: nil,
-          database: %Database{
-            sid: sid,
-            health: :passing,
-            instances: [
-              %Instance{
-                sid: sid,
-                system_replication: "Primary",
-                system_replication_status: "ACTIVE",
-                instance_number: instance_number,
-                features: features,
-                host_id: host_id,
-                health: :passing
-              }
-            ]
-          }
-        }
-      )
-    end
-
-    test "should add a database instance to an existing Database" do
-      sap_system_id = Faker.UUID.v4()
-      sid = Faker.StarWars.planet()
-      tenant = Faker.Beer.style()
-      instance_number = "00"
-      features = Faker.Pokemon.name()
-      host_id = Faker.UUID.v4()
-
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          tenant: tenant,
-          instance_number: "10"
-        )
-      ]
-
-      assert_events_and_state(
-        initial_events,
-        build(
-          :register_database_instance_command,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          tenant: tenant,
-          instance_number: instance_number,
-          features: features,
-          host_id: host_id,
-          health: :passing
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          tenant: tenant,
-          instance_number: instance_number,
-          features: features,
-          host_id: host_id,
-          health: :passing
-        ),
-        fn state ->
-          assert %SapSystem{
-                   database: %Database{
-                     instances: [
-                       %Instance{
-                         sid: ^sid,
-                         instance_number: ^instance_number,
-                         features: ^features,
-                         host_id: ^host_id,
-                         health: :passing
-                       }
-                       | _
-                     ]
-                   }
-                 } = state
-        end
-      )
-    end
-
-    test "should not add a database instance if the database instance was already registered" do
-      database_registered_event = build(:database_registered_event)
-
-      database_instance_registered_event =
-        build(
-          :database_instance_registered_event,
-          sap_system_id: database_registered_event.sap_system_id
-        )
-
-      initial_events = [
-        database_registered_event,
-        database_instance_registered_event
-      ]
-
-      assert_events(
-        initial_events,
-        build(
-          :register_database_instance_command,
-          sap_system_id: database_registered_event.sap_system_id,
-          sid: database_instance_registered_event.sid,
-          tenant: database_instance_registered_event.tenant,
-          instance_number: database_instance_registered_event.instance_number,
-          features: database_instance_registered_event.features,
-          host_id: database_instance_registered_event.host_id,
-          system_replication: database_instance_registered_event.system_replication,
-          system_replication_status: database_instance_registered_event.system_replication_status,
-          health: :passing
-        ),
-        []
-      )
-    end
-
-    test "should change the system replication of a database instance" do
-      database_registered_event = build(:database_registered_event)
-
-      database_instance_registered_event =
-        build(
-          :database_instance_registered_event,
-          sap_system_id: database_registered_event.sap_system_id,
-          system_replication: "Secondary",
-          system_replication_status: ""
-        )
-
-      initial_events = [
-        database_registered_event,
-        database_instance_registered_event
-      ]
-
-      assert_events(
-        initial_events,
-        build(
-          :register_database_instance_command,
-          sap_system_id: database_registered_event.sap_system_id,
-          sid: database_instance_registered_event.sid,
-          tenant: database_instance_registered_event.tenant,
-          instance_number: database_instance_registered_event.instance_number,
-          features: database_instance_registered_event.features,
-          host_id: database_instance_registered_event.host_id,
-          system_replication: "Primary",
-          system_replication_status: "ACTIVE",
-          health: :passing
-        ),
-        %DatabaseInstanceSystemReplicationChanged{
-          sap_system_id: database_registered_event.sap_system_id,
-          host_id: database_instance_registered_event.host_id,
-          instance_number: database_instance_registered_event.instance_number,
-          system_replication: "Primary",
-          system_replication_status: "ACTIVE"
-        }
-      )
-    end
-
     test "should register a SAP System and add an application instance when a MESSAGESERVER instance is already present and a new ABAP instance is added" do
       sap_system_id = Faker.UUID.v4()
       sid = Faker.StarWars.planet()
@@ -361,17 +48,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       ensa_version = EnsaVersion.ensa1()
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          tenant: tenant
-        ),
         build(:application_instance_registered_event,
           sap_system_id: sap_system_id,
           sid: sid,
@@ -423,22 +99,19 @@ defmodule Trento.SapSystems.SapSystemTest do
           assert %SapSystem{
                    sid: ^sid,
                    ensa_version: ^ensa_version,
-                   application: %Application{
-                     sid: ^sid,
-                     instances: [
-                       %Instance{
-                         sid: ^sid,
-                         instance_number: "10",
-                         features: "ABAP",
-                         host_id: ^host_id,
-                         health: :passing,
-                         absent_at: nil
-                       },
-                       %Instance{
-                         features: "MESSAGESERVER"
-                       }
-                     ]
-                   }
+                   instances: [
+                     %Instance{
+                       sid: ^sid,
+                       instance_number: "10",
+                       features: "ABAP",
+                       host_id: ^host_id,
+                       health: :passing,
+                       absent_at: nil
+                     },
+                     %Instance{
+                       features: "MESSAGESERVER"
+                     }
+                   ]
                  } = state
         end
       )
@@ -460,16 +133,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       cluster_id = Faker.UUID.v4()
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
         build(:application_instance_registered_event,
           sap_system_id: sap_system_id,
           sid: sid,
@@ -523,16 +186,14 @@ defmodule Trento.SapSystems.SapSystemTest do
         ],
         fn state ->
           assert %SapSystem{
-                   application: %Application{
-                     instances: [
-                       %Instance{
-                         sid: ^sid,
-                         instance_number: ^instance_number,
-                         host_id: ^new_host_id
-                       }
-                       | _
-                     ]
-                   }
+                   instances: [
+                     %Instance{
+                       sid: ^sid,
+                       instance_number: ^instance_number,
+                       host_id: ^new_host_id
+                     }
+                     | _
+                   ]
                  } = state
         end
       )
@@ -554,16 +215,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       cluster_id = Faker.UUID.v4()
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
         build(:application_instance_registered_event,
           sap_system_id: sap_system_id,
           sid: sid,
@@ -628,16 +279,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       ensa_version = EnsaVersion.ensa1()
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
         build(:application_instance_registered_event,
           sap_system_id: sap_system_id,
           sid: sid,
@@ -697,17 +338,15 @@ defmodule Trento.SapSystems.SapSystemTest do
         ],
         fn state ->
           assert %SapSystem{
-                   application: %Application{
-                     instances: [
-                       %Instance{
-                         sid: ^sid,
-                         instance_number: ^instance_number,
-                         host_id: ^new_host_id
-                       },
-                       _,
-                       _
-                     ]
-                   }
+                   instances: [
+                     %Instance{
+                       sid: ^sid,
+                       instance_number: ^instance_number,
+                       host_id: ^new_host_id
+                     },
+                     _,
+                     _
+                   ]
                  } = state
         end
       )
@@ -727,16 +366,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       ensa_version = EnsaVersion.ensa1()
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
         build(:application_instance_registered_event,
           sap_system_id: sap_system_id,
           sid: sid,
@@ -782,16 +411,14 @@ defmodule Trento.SapSystems.SapSystemTest do
         [],
         fn state ->
           assert %SapSystem{
-                   application: %Application{
-                     instances: [
-                       %Instance{
-                         sid: ^sid,
-                         instance_number: ^instance_number,
-                         host_id: ^old_host_id
-                       },
-                       _
-                     ]
-                   }
+                   instances: [
+                     %Instance{
+                       sid: ^sid,
+                       instance_number: ^instance_number,
+                       host_id: ^old_host_id
+                     },
+                     _
+                   ]
                  } = state
         end
       )
@@ -805,16 +432,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       host_id = Faker.UUID.v4()
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
         build(:application_instance_registered_event,
           sap_system_id: sap_system_id,
           sid: sid,
@@ -854,10 +471,7 @@ defmodule Trento.SapSystems.SapSystemTest do
         fn state ->
           assert %SapSystem{
                    sid: ^sid,
-                   ensa_version: ^ensa_version,
-                   application: %Application{
-                     sid: ^sid
-                   }
+                   ensa_version: ^ensa_version
                  } = state
         end
       )
@@ -871,16 +485,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       host_id = Faker.UUID.v4()
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
         build(:application_instance_registered_event,
           sap_system_id: sap_system_id,
           sid: sid,
@@ -915,10 +519,7 @@ defmodule Trento.SapSystems.SapSystemTest do
         fn state ->
           assert %SapSystem{
                    sid: ^sid,
-                   ensa_version: ^ensa_version,
-                   application: %Application{
-                     sid: ^sid
-                   }
+                   ensa_version: ^ensa_version
                  } = state
         end
       )
@@ -932,16 +533,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       host_id = Faker.UUID.v4()
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
         build(:application_instance_registered_event,
           sap_system_id: sap_system_id,
           sid: sid,
@@ -976,10 +567,7 @@ defmodule Trento.SapSystems.SapSystemTest do
         fn state ->
           assert %SapSystem{
                    sid: ^sid,
-                   ensa_version: ^ensa_version,
-                   application: %Application{
-                     sid: ^sid
-                   }
+                   ensa_version: ^ensa_version
                  } = state
         end
       )
@@ -998,17 +586,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       ensa_version = EnsaVersion.ensa1()
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          tenant: tenant
-        ),
         build(:application_instance_registered_event,
           sap_system_id: sap_system_id,
           sid: sid,
@@ -1059,21 +636,18 @@ defmodule Trento.SapSystems.SapSystemTest do
         fn state ->
           assert %SapSystem{
                    sid: ^sid,
-                   application: %Application{
-                     sid: ^sid,
-                     instances: [
-                       %Instance{
-                         sid: ^sid,
-                         instance_number: "00",
-                         features: "MESSAGESERVER",
-                         host_id: ^host_id,
-                         health: :passing
-                       },
-                       %Instance{
-                         features: "ABAP"
-                       }
-                     ]
-                   }
+                   instances: [
+                     %Instance{
+                       sid: ^sid,
+                       instance_number: "00",
+                       features: "MESSAGESERVER",
+                       host_id: ^host_id,
+                       health: :passing
+                     },
+                     %Instance{
+                       features: "ABAP"
+                     }
+                   ]
                  } = state
         end
       )
@@ -1091,19 +665,7 @@ defmodule Trento.SapSystems.SapSystemTest do
       host_id = Faker.UUID.v4()
       ensa_version = EnsaVersion.ensa1()
 
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          tenant: tenant
-        )
-      ]
+      initial_events = []
 
       assert_events_and_state(
         initial_events,
@@ -1139,18 +701,15 @@ defmodule Trento.SapSystems.SapSystemTest do
         fn state ->
           assert %SapSystem{
                    sid: nil,
-                   application: %Application{
-                     sid: ^sid,
-                     instances: [
-                       %Instance{
-                         sid: ^sid,
-                         instance_number: "00",
-                         features: "ABAP",
-                         host_id: ^host_id,
-                         health: :passing
-                       }
-                     ]
-                   }
+                   instances: [
+                     %Instance{
+                       sid: ^sid,
+                       instance_number: "00",
+                       features: "ABAP",
+                       host_id: ^host_id,
+                       health: :passing
+                     }
+                   ]
                  } = state
         end
       )
@@ -1168,19 +727,7 @@ defmodule Trento.SapSystems.SapSystemTest do
       host_id = Faker.UUID.v4()
       ensa_version = EnsaVersion.ensa1()
 
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          tenant: tenant
-        )
-      ]
+      initial_events = []
 
       assert_events_and_state(
         initial_events,
@@ -1216,18 +763,15 @@ defmodule Trento.SapSystems.SapSystemTest do
         fn state ->
           assert %SapSystem{
                    sid: nil,
-                   application: %Application{
-                     sid: ^sid,
-                     instances: [
-                       %Instance{
-                         sid: ^sid,
-                         instance_number: "00",
-                         features: "MESSAGESERVER",
-                         host_id: ^host_id,
-                         health: :passing
-                       }
-                     ]
-                   }
+                   instances: [
+                     %Instance{
+                       sid: ^sid,
+                       instance_number: "00",
+                       features: "MESSAGESERVER",
+                       host_id: ^host_id,
+                       health: :passing
+                     }
+                   ]
                  } = state
         end
       )
@@ -1238,8 +782,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       sid = Faker.StarWars.planet()
 
       initial_events = [
-        build(:database_registered_event, sap_system_id: sap_system_id, sid: sid),
-        build(:database_instance_registered_event, sap_system_id: sap_system_id, sid: sid),
         build(:application_instance_registered_event, sap_system_id: sap_system_id, sid: sid),
         build(:sap_system_registered_event, sap_system_id: sap_system_id, sid: sid)
       ]
@@ -1274,19 +816,17 @@ defmodule Trento.SapSystems.SapSystemTest do
         ),
         fn state ->
           assert %SapSystem{
-                   application: %Application{
-                     sid: ^sid,
-                     instances: [
-                       %Instance{
-                         sid: ^sid,
-                         instance_number: ^new_instance_number,
-                         features: ^new_instance_features,
-                         host_id: ^new_instance_host_id,
-                         health: :passing
-                       }
-                       | _
-                     ]
-                   }
+                   sid: ^sid,
+                   instances: [
+                     %Instance{
+                       sid: ^sid,
+                       instance_number: ^new_instance_number,
+                       features: ^new_instance_features,
+                       host_id: ^new_instance_host_id,
+                       health: :passing
+                     }
+                     | _
+                   ]
                  } = state
         end
       )
@@ -1299,10 +839,7 @@ defmodule Trento.SapSystems.SapSystemTest do
         build(:application_instance_registered_event, sap_system_id: sap_system_id)
 
       initial_events = [
-        build(:database_registered_event, sap_system_id: sap_system_id),
-        build(:database_instance_registered_event, sap_system_id: sap_system_id),
-        application_instance_registered_event,
-        build(:sap_system_registered_event, sap_system_id: sap_system_id)
+        application_instance_registered_event
       ]
 
       assert_events(
@@ -1324,197 +861,6 @@ defmodule Trento.SapSystems.SapSystemTest do
   end
 
   describe "SAP System health" do
-    test "should change the health of a Database when a new Database instance is registered" do
-      sap_system_id = Faker.UUID.v4()
-      sid = Faker.StarWars.planet()
-      tenant = Faker.Beer.style()
-      instance_number = "00"
-      features = Faker.Pokemon.name()
-      host_id = Faker.UUID.v4()
-
-      initial_events = [
-        build(:database_registered_event, sap_system_id: sap_system_id),
-        build(:database_instance_registered_event, sap_system_id: sap_system_id)
-      ]
-
-      assert_events_and_state(
-        initial_events,
-        build(
-          :register_database_instance_command,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          tenant: tenant,
-          instance_number: instance_number,
-          features: features,
-          host_id: host_id,
-          health: :critical
-        ),
-        [
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            sid: sid,
-            tenant: tenant,
-            instance_number: instance_number,
-            features: features,
-            host_id: host_id,
-            health: :critical
-          ),
-          %DatabaseHealthChanged{
-            sap_system_id: sap_system_id,
-            health: :critical
-          }
-        ],
-        fn state ->
-          %SapSystem{
-            database: %Database{
-              health: :critical,
-              instances: [
-                %Instance{
-                  health: :critical
-                },
-                %Instance{
-                  health: :passing
-                }
-              ]
-            }
-          } = state
-        end
-      )
-    end
-
-    test "should change the health of a Database when a Database instance has changed the health status" do
-      sap_system_id = Faker.UUID.v4()
-      host_id = Faker.UUID.v4()
-      instance_number = "00"
-
-      database_instance_registered_event =
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: host_id,
-          instance_number: instance_number
-        )
-
-      initial_events = [
-        build(:database_registered_event, sap_system_id: sap_system_id),
-        database_instance_registered_event
-      ]
-
-      assert_events_and_state(
-        initial_events,
-        build(
-          :register_database_instance_command,
-          sap_system_id: sap_system_id,
-          sid: database_instance_registered_event.sid,
-          tenant: database_instance_registered_event.tenant,
-          instance_number: instance_number,
-          features: database_instance_registered_event.features,
-          host_id: host_id,
-          health: :critical
-        ),
-        [
-          %DatabaseInstanceHealthChanged{
-            sap_system_id: sap_system_id,
-            instance_number: instance_number,
-            host_id: host_id,
-            health: :critical
-          },
-          %DatabaseHealthChanged{
-            sap_system_id: sap_system_id,
-            health: :critical
-          }
-        ],
-        fn state ->
-          assert %SapSystem{
-                   database: %Database{
-                     health: :critical,
-                     instances: [
-                       %Instance{
-                         instance_number: ^instance_number,
-                         host_id: ^host_id,
-                         health: :critical
-                       }
-                     ]
-                   }
-                 } = state
-        end
-      )
-    end
-
-    test "should not change the health of a Database if no instance has changed the health status" do
-      sap_system_id = Faker.UUID.v4()
-
-      new_instance_number = "20"
-      new_instance_features = Faker.Pokemon.name()
-      new_instance_host_id = Faker.UUID.v4()
-
-      database_instance_registered_event =
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          health: :warning
-        )
-
-      initial_events = [
-        build(:database_registered_event, sap_system_id: sap_system_id, health: :warning),
-        database_instance_registered_event
-      ]
-
-      assert_events_and_state(
-        initial_events,
-        [
-          build(
-            :register_database_instance_command,
-            sap_system_id: sap_system_id,
-            sid: database_instance_registered_event.sid,
-            tenant: database_instance_registered_event.tenant,
-            instance_number: database_instance_registered_event.instance_number,
-            features: database_instance_registered_event.features,
-            host_id: database_instance_registered_event.host_id,
-            health: :warning
-          ),
-          build(
-            :register_database_instance_command,
-            sap_system_id: sap_system_id,
-            sid: database_instance_registered_event.sid,
-            tenant: database_instance_registered_event.tenant,
-            instance_number: new_instance_number,
-            features: new_instance_features,
-            host_id: new_instance_host_id,
-            health: :warning
-          )
-        ],
-        [
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            sid: database_instance_registered_event.sid,
-            tenant: database_instance_registered_event.tenant,
-            instance_number: new_instance_number,
-            features: new_instance_features,
-            host_id: new_instance_host_id,
-            health: :warning
-          )
-        ],
-        fn state ->
-          assert %SapSystem{
-                   database: %Database{
-                     health: :warning,
-                     instances: [
-                       %Instance{
-                         health: :warning
-                       },
-                       %Instance{
-                         health: :warning
-                       }
-                     ]
-                   }
-                 } = state
-        end
-      )
-    end
-
     test "should change the health of a SAP System when a new Application instance is registered" do
       sap_system_id = Faker.UUID.v4()
       sid = Faker.StarWars.planet()
@@ -1526,17 +872,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       ensa_version = EnsaVersion.ensa1()
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          tenant: tenant
-        ),
         build(:application_instance_registered_event,
           sap_system_id: sap_system_id,
           sid: sid,
@@ -1581,16 +916,14 @@ defmodule Trento.SapSystems.SapSystemTest do
           assert %SapSystem{
                    health: :critical,
                    ensa_version: ^ensa_version,
-                   application: %Application{
-                     instances: [
-                       %Instance{
-                         health: :critical
-                       },
-                       %Instance{
-                         health: :passing
-                       }
-                     ]
-                   }
+                   instances: [
+                     %Instance{
+                       health: :critical
+                     },
+                     %Instance{
+                       health: :passing
+                     }
+                   ]
                  } = state
         end
       )
@@ -1606,8 +939,6 @@ defmodule Trento.SapSystems.SapSystemTest do
         build(:sap_system_registered_event, sap_system_id: sap_system_id)
 
       initial_events = [
-        build(:database_registered_event, sap_system_id: sap_system_id),
-        build(:database_instance_registered_event, sap_system_id: sap_system_id),
         application_instance_registered,
         sap_system_registered_event
       ]
@@ -1640,13 +971,11 @@ defmodule Trento.SapSystems.SapSystemTest do
         fn state ->
           assert %SapSystem{
                    health: :critical,
-                   application: %Application{
-                     instances: [
-                       %Instance{
-                         health: :critical
-                       }
-                     ]
-                   }
+                   instances: [
+                     %Instance{
+                       health: :critical
+                     }
+                   ]
                  } = state
         end
       )
@@ -1670,12 +999,6 @@ defmodule Trento.SapSystems.SapSystemTest do
         build(:sap_system_registered_event, sap_system_id: sap_system_id, health: :warning)
 
       initial_events = [
-        build(:database_registered_event, sap_system_id: sap_system_id),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          health: :warning
-        ),
         application_instance_registered_event,
         sap_system_registered_event
       ]
@@ -1720,75 +1043,17 @@ defmodule Trento.SapSystems.SapSystemTest do
         fn state ->
           assert %SapSystem{
                    health: :warning,
-                   application: %Application{
-                     instances: [
-                       %Instance{
-                         health: :warning
-                       },
-                       %Instance{
-                         health: :warning
-                       }
-                     ]
-                   }
+                   instances: [
+                     %Instance{
+                       health: :warning
+                     },
+                     %Instance{
+                       health: :warning
+                     }
+                   ]
                  } = state
         end
       )
-    end
-
-    test "should change the health of a SAP System when the Database has changed the health status" do
-      sap_system_id = Faker.UUID.v4()
-
-      new_instance_number = "20"
-      new_instance_features = Faker.Pokemon.name()
-      new_instance_host_id = Faker.UUID.v4()
-
-      initial_events = [
-        build(:database_registered_event, sap_system_id: sap_system_id),
-        database_instance_registered_event =
-          build(:database_instance_registered_event, sap_system_id: sap_system_id),
-        build(:application_instance_registered_event, sap_system_id: sap_system_id),
-        build(:sap_system_registered_event, sap_system_id: sap_system_id)
-      ]
-
-      assert_events_and_state(
-        initial_events,
-        build(
-          :register_database_instance_command,
-          sap_system_id: sap_system_id,
-          sid: database_instance_registered_event.sid,
-          tenant: database_instance_registered_event.tenant,
-          instance_number: new_instance_number,
-          features: new_instance_features,
-          host_id: new_instance_host_id,
-          health: :warning
-        ),
-        [
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            sid: database_instance_registered_event.sid,
-            tenant: database_instance_registered_event.tenant,
-            instance_number: new_instance_number,
-            features: new_instance_features,
-            host_id: new_instance_host_id,
-            health: :warning
-          ),
-          %DatabaseHealthChanged{
-            sap_system_id: sap_system_id,
-            health: :warning
-          },
-          %SapSystemHealthChanged{
-            sap_system_id: sap_system_id,
-            health: :warning
-          }
-        ],
-        fn state ->
-          assert %SapSystem{health: :warning} = state
-        end
-      )
-    end
-
-    test "should update the SAP system if some of the fields have been changed" do
     end
   end
 
@@ -1804,15 +1069,10 @@ defmodule Trento.SapSystems.SapSystemTest do
       sap_system_id = UUID.uuid4()
       sid = UUID.uuid4()
 
-      database_instance_registered_event =
-        build(:database_instance_registered_event, sap_system_id: sap_system_id, sid: sid)
-
       application_instance_registered_event =
         build(:application_instance_registered_event, sap_system_id: sap_system_id, sid: sid)
 
       initial_events = [
-        build(:database_registered_event, sap_system_id: sap_system_id, sid: sid),
-        database_instance_registered_event,
         application_instance_registered_event,
         %{ensa_version: ensa_version} =
           build(:sap_system_registered_event, sap_system_id: sap_system_id, sid: sid)
@@ -1828,36 +1088,17 @@ defmodule Trento.SapSystems.SapSystemTest do
             sid: sid,
             health: :passing,
             ensa_version: ensa_version,
-            application: %Application{
-              sid: sid,
-              instances: [
-                %Instance{
-                  sid: sid,
-                  instance_number: application_instance_registered_event.instance_number,
-                  health: application_instance_registered_event.health,
-                  features: application_instance_registered_event.features,
-                  host_id: application_instance_registered_event.host_id,
-                  system_replication: nil,
-                  system_replication_status: nil
-                }
-              ]
-            },
-            database: %Database{
-              sid: sid,
-              health: :passing,
-              instances: [
-                %Instance{
-                  sid: sid,
-                  instance_number: database_instance_registered_event.instance_number,
-                  health: database_instance_registered_event.health,
-                  features: database_instance_registered_event.features,
-                  host_id: database_instance_registered_event.host_id,
-                  system_replication: database_instance_registered_event.system_replication,
-                  system_replication_status:
-                    database_instance_registered_event.system_replication_status
-                }
-              ]
-            },
+            instances: [
+              %Instance{
+                sid: sid,
+                instance_number: application_instance_registered_event.instance_number,
+                health: application_instance_registered_event.health,
+                features: application_instance_registered_event.features,
+                host_id: application_instance_registered_event.host_id,
+                system_replication: nil,
+                system_replication_status: nil
+              }
+            ],
             rolling_up: false
           }
         },
@@ -1872,8 +1113,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       sid = UUID.uuid4()
 
       initial_events = [
-        build(:database_registered_event, sap_system_id: sap_system_id, sid: sid),
-        build(:database_instance_registered_event, sap_system_id: sap_system_id, sid: sid),
         build(:application_instance_registered_event, sap_system_id: sap_system_id, sid: sid),
         build(:sap_system_registered_event, sap_system_id: sap_system_id, sid: sid),
         %SapSystemRollUpRequested{
@@ -1884,17 +1123,10 @@ defmodule Trento.SapSystems.SapSystemTest do
 
       assert_error(
         initial_events,
-        RegisterDatabaseInstance.new!(%{
-          sap_system_id: sap_system_id,
-          sid: Faker.StarWars.planet(),
-          tenant: Faker.UUID.v4(),
-          host_id: Faker.UUID.v4(),
-          instance_number: "00",
-          features: Faker.Pokemon.name(),
-          http_port: 8080,
-          https_port: 8443,
-          health: :passing
-        }),
+        build(
+          :register_application_instance_command,
+          sap_system_id: sap_system_id
+        ),
         {:error, :sap_system_rolling_up}
       )
 
@@ -1941,45 +1173,18 @@ defmodule Trento.SapSystems.SapSystemTest do
   end
 
   describe "tombstoning" do
-    test "should tombstone a deregistered SAP system when no application and no database instances are left" do
+    test "should tombstone a deregistered SAP system when no application instances are left" do
       sap_system_id = UUID.uuid4()
-      host_id = UUID.uuid4()
-      secondary_database_host_id = UUID.uuid4()
       deregistered_at = DateTime.utc_now()
-
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
 
       message_server_host_id = UUID.uuid4()
       message_server_instance_number = "00"
       abap_host_id = UUID.uuid4()
       abap_instance_number = "01"
 
-      db_sid = fake_sid()
       application_sid = fake_sid()
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: host_id,
-          instance_number: db_instance_number_1,
-          system_replication: "Primary",
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_2,
-          system_replication: "Secondary",
-          sid: db_sid
-        ),
         build(
           :application_instance_registered_event,
           sap_system_id: sap_system_id,
@@ -2006,18 +1211,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       assert_events_and_state(
         initial_events,
         [
-          %DeregisterDatabaseInstance{
-            sap_system_id: sap_system_id,
-            host_id: host_id,
-            instance_number: db_instance_number_1,
-            deregistered_at: deregistered_at
-          },
-          %DeregisterDatabaseInstance{
-            sap_system_id: sap_system_id,
-            host_id: secondary_database_host_id,
-            instance_number: db_instance_number_2,
-            deregistered_at: deregistered_at
-          },
           %DeregisterApplicationInstance{
             sap_system_id: sap_system_id,
             host_id: message_server_host_id,
@@ -2032,30 +1225,14 @@ defmodule Trento.SapSystems.SapSystemTest do
           }
         ],
         [
-          %DatabaseInstanceDeregistered{
-            sap_system_id: sap_system_id,
-            host_id: host_id,
-            instance_number: db_instance_number_1,
-            deregistered_at: deregistered_at
-          },
-          %DatabaseDeregistered{
-            sap_system_id: sap_system_id,
-            deregistered_at: deregistered_at
-          },
-          %SapSystemDeregistered{
-            sap_system_id: sap_system_id,
-            deregistered_at: deregistered_at
-          },
-          %DatabaseInstanceDeregistered{
-            sap_system_id: sap_system_id,
-            host_id: secondary_database_host_id,
-            instance_number: db_instance_number_2,
-            deregistered_at: deregistered_at
-          },
           %ApplicationInstanceDeregistered{
             sap_system_id: sap_system_id,
             host_id: message_server_host_id,
             instance_number: message_server_instance_number,
+            deregistered_at: deregistered_at
+          },
+          %SapSystemDeregistered{
+            sap_system_id: sap_system_id,
             deregistered_at: deregistered_at
           },
           %ApplicationInstanceDeregistered{
@@ -2070,14 +1247,7 @@ defmodule Trento.SapSystems.SapSystemTest do
         ],
         fn sap_system ->
           assert %SapSystem{
-                   database: %Database{
-                     deregistered_at: ^deregistered_at,
-                     sid: ^db_sid,
-                     instances: []
-                   },
-                   application: %Application{
-                     instances: []
-                   },
+                   instances: [],
                    deregistered_at: ^deregistered_at,
                    sid: ^application_sid
                  } = sap_system
@@ -2087,653 +1257,13 @@ defmodule Trento.SapSystems.SapSystemTest do
   end
 
   describe "deregistration" do
-    test "should not restore a deregistered database when the registering database instance has Secondary role" do
-      sap_system_id = UUID.uuid4()
-
-      primary_database_host_id = UUID.uuid4()
-      secondary_database_host_id = UUID.uuid4()
-
-      deregistered_at = DateTime.utc_now()
-
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
-
-      db_sid = fake_sid()
-      application_sid = fake_sid()
-
-      message_server_host_id = UUID.uuid4()
-      message_server_instance_number = "00"
-      abap_host_id = UUID.uuid4()
-      abap_instance_number = "01"
-
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          sid: db_sid,
-          instance_number: db_instance_number_1,
-          system_replication: "Primary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_2,
-          system_replication: "Secondary",
-          sid: db_sid
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "MESSAGESERVER|ENQUE",
-          host_id: message_server_host_id,
-          instance_number: message_server_instance_number,
-          sid: application_sid
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "ABAP|GATEWAY|ICMAN|IGS",
-          host_id: abap_host_id,
-          instance_number: abap_instance_number,
-          sid: application_sid
-        ),
-        build(
-          :sap_system_registered_event,
-          sap_system_id: sap_system_id,
-          sid: application_sid
-        ),
-        build(:database_instance_deregistered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          instance_number: db_instance_number_1,
-          deregistered_at: deregistered_at
-        ),
-        build(:database_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        ),
-        build(:sap_system_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        )
-      ]
-
-      command =
-        build(:register_database_instance_command,
-          system_replication: "Secondary",
-          sid: db_sid,
-          sap_system_id: sap_system_id
-        )
-
-      assert_error(
-        initial_events,
-        command,
-        {:error, :sap_system_not_registered}
-      )
-    end
-
-    test "should restore a deregistered database when the registering database instance has system replication disabled, with database instance leftovers" do
-      sap_system_id = UUID.uuid4()
-
-      primary_database_host_id = UUID.uuid4()
-      secondary_database_host_id = UUID.uuid4()
-
-      deregistered_at = DateTime.utc_now()
-
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
-
-      db_sid = fake_sid()
-      application_sid = fake_sid()
-
-      message_server_host_id = UUID.uuid4()
-      message_server_instance_number = "00"
-      abap_host_id = UUID.uuid4()
-      abap_instance_number = "01"
-
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          sid: db_sid,
-          instance_number: db_instance_number_1,
-          system_replication: "Primary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_2,
-          system_replication: "Secondary",
-          sid: db_sid
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "MESSAGESERVER|ENQUE",
-          host_id: message_server_host_id,
-          instance_number: message_server_instance_number,
-          sid: application_sid
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "ABAP|GATEWAY|ICMAN|IGS",
-          host_id: abap_host_id,
-          instance_number: abap_instance_number,
-          sid: application_sid
-        ),
-        build(
-          :sap_system_registered_event,
-          sap_system_id: sap_system_id,
-          sid: application_sid
-        ),
-        build(:database_instance_deregistered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          instance_number: db_instance_number_1,
-          deregistered_at: deregistered_at
-        ),
-        build(:database_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        ),
-        build(:sap_system_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        )
-      ]
-
-      %{features: features, instance_number: instance_number, health: health} =
-        command =
-        build(:register_database_instance_command,
-          system_replication: nil,
-          sid: db_sid,
-          sap_system_id: sap_system_id
-        )
-
-      assert_events_and_state(
-        initial_events,
-        command,
-        [
-          %DatabaseInstanceRegistered{
-            sap_system_id: sap_system_id,
-            sid: db_sid,
-            tenant: command.tenant,
-            instance_number: command.instance_number,
-            instance_hostname: command.instance_hostname,
-            features: command.features,
-            http_port: command.http_port,
-            https_port: command.https_port,
-            start_priority: command.start_priority,
-            host_id: command.host_id,
-            system_replication: command.system_replication,
-            system_replication_status: command.system_replication_status,
-            health: command.health
-          },
-          %DatabaseRestored{
-            sap_system_id: sap_system_id,
-            health: command.health
-          }
-        ],
-        fn sap_system ->
-          assert %SapSystem{
-                   deregistered_at: ^deregistered_at,
-                   database: %Database{
-                     deregistered_at: nil,
-                     sid: ^db_sid,
-                     instances: [
-                       %Instance{
-                         sid: ^db_sid,
-                         instance_number: ^instance_number,
-                         features: ^features,
-                         health: ^health
-                       },
-                       %Instance{}
-                     ]
-                   }
-                 } = sap_system
-        end
-      )
-    end
-
-    test "should restore a deregistered database when the registering database instance has system replication disabled, without database instance leftovers" do
-      sap_system_id = UUID.uuid4()
-
-      primary_database_host_id = UUID.uuid4()
-      secondary_database_host_id = UUID.uuid4()
-
-      deregistered_at = DateTime.utc_now()
-
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
-
-      db_sid = fake_sid()
-      application_sid = fake_sid()
-
-      message_server_host_id = UUID.uuid4()
-      message_server_instance_number = "00"
-      abap_host_id = UUID.uuid4()
-      abap_instance_number = "01"
-
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          sid: db_sid,
-          instance_number: db_instance_number_1,
-          system_replication: "Primary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_2,
-          system_replication: "Secondary",
-          sid: db_sid
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "MESSAGESERVER|ENQUE",
-          host_id: message_server_host_id,
-          instance_number: message_server_instance_number,
-          sid: application_sid
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "ABAP|GATEWAY|ICMAN|IGS",
-          host_id: abap_host_id,
-          instance_number: abap_instance_number,
-          sid: application_sid
-        ),
-        build(
-          :sap_system_registered_event,
-          sap_system_id: sap_system_id,
-          sid: application_sid
-        ),
-        build(:database_instance_deregistered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          instance_number: db_instance_number_1,
-          deregistered_at: deregistered_at
-        ),
-        build(:database_instance_deregistered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_2,
-          deregistered_at: deregistered_at
-        ),
-        build(:database_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        ),
-        build(:sap_system_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        )
-      ]
-
-      %{features: features, instance_number: instance_number, health: health} =
-        command =
-        build(:register_database_instance_command,
-          system_replication: nil,
-          sid: db_sid,
-          sap_system_id: sap_system_id
-        )
-
-      assert_events_and_state(
-        initial_events,
-        command,
-        [
-          %DatabaseInstanceRegistered{
-            sap_system_id: sap_system_id,
-            sid: db_sid,
-            tenant: command.tenant,
-            instance_number: command.instance_number,
-            instance_hostname: command.instance_hostname,
-            features: command.features,
-            http_port: command.http_port,
-            https_port: command.https_port,
-            start_priority: command.start_priority,
-            host_id: command.host_id,
-            system_replication: command.system_replication,
-            system_replication_status: command.system_replication_status,
-            health: command.health
-          },
-          %DatabaseRestored{
-            sap_system_id: sap_system_id,
-            health: command.health
-          }
-        ],
-        fn sap_system ->
-          assert Kernel.length(sap_system.database.instances) == 1
-
-          assert %SapSystem{
-                   deregistered_at: ^deregistered_at,
-                   database: %Database{
-                     deregistered_at: nil,
-                     sid: ^db_sid,
-                     instances: [
-                       %Instance{
-                         sid: ^db_sid,
-                         instance_number: ^instance_number,
-                         features: ^features,
-                         health: ^health,
-                         system_replication: nil
-                       }
-                     ]
-                   }
-                 } = sap_system
-        end
-      )
-    end
-
-    test "should restore a deregistered database when the registering database instance is a primary, without database instance leftovers" do
-      sap_system_id = UUID.uuid4()
-
-      primary_database_host_id = UUID.uuid4()
-      secondary_database_host_id = UUID.uuid4()
-
-      deregistered_at = DateTime.utc_now()
-
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
-
-      db_sid = fake_sid()
-      application_sid = fake_sid()
-
-      message_server_host_id = UUID.uuid4()
-      message_server_instance_number = "00"
-      abap_host_id = UUID.uuid4()
-      abap_instance_number = "01"
-
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          sid: db_sid,
-          instance_number: db_instance_number_1,
-          system_replication: "Primary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_2,
-          system_replication: "Secondary",
-          sid: db_sid
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "MESSAGESERVER|ENQUE",
-          host_id: message_server_host_id,
-          instance_number: message_server_instance_number,
-          sid: application_sid
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "ABAP|GATEWAY|ICMAN|IGS",
-          host_id: abap_host_id,
-          instance_number: abap_instance_number,
-          sid: application_sid
-        ),
-        build(
-          :sap_system_registered_event,
-          sap_system_id: sap_system_id,
-          sid: application_sid
-        ),
-        build(:database_instance_deregistered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          instance_number: db_instance_number_1,
-          deregistered_at: deregistered_at
-        ),
-        build(:database_instance_deregistered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_2,
-          deregistered_at: deregistered_at
-        ),
-        build(:database_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        ),
-        build(:sap_system_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        )
-      ]
-
-      %{features: features, instance_number: instance_number, health: health} =
-        command =
-        build(:register_database_instance_command,
-          system_replication: "Primary",
-          sid: db_sid,
-          sap_system_id: sap_system_id
-        )
-
-      assert_events_and_state(
-        initial_events,
-        command,
-        [
-          %DatabaseInstanceRegistered{
-            sap_system_id: sap_system_id,
-            sid: db_sid,
-            tenant: command.tenant,
-            instance_number: command.instance_number,
-            instance_hostname: command.instance_hostname,
-            features: command.features,
-            http_port: command.http_port,
-            https_port: command.https_port,
-            start_priority: command.start_priority,
-            host_id: command.host_id,
-            system_replication: command.system_replication,
-            system_replication_status: command.system_replication_status,
-            health: command.health
-          },
-          %DatabaseRestored{
-            sap_system_id: sap_system_id,
-            health: command.health
-          }
-        ],
-        fn sap_system ->
-          assert Kernel.length(sap_system.database.instances) == 1
-
-          assert %SapSystem{
-                   deregistered_at: ^deregistered_at,
-                   database: %Database{
-                     deregistered_at: nil,
-                     sid: ^db_sid,
-                     instances: [
-                       %Instance{
-                         sid: ^db_sid,
-                         instance_number: ^instance_number,
-                         features: ^features,
-                         health: ^health,
-                         system_replication: "Primary"
-                       }
-                     ]
-                   }
-                 } = sap_system
-        end
-      )
-    end
-
-    test "should restore a deregistered database when the registering database instance is a primary, with database instance leftovers" do
-      sap_system_id = UUID.uuid4()
-
-      primary_database_host_id = UUID.uuid4()
-      secondary_database_host_id = UUID.uuid4()
-
-      deregistered_at = DateTime.utc_now()
-
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
-
-      db_sid = fake_sid()
-      application_sid = fake_sid()
-
-      message_server_host_id = UUID.uuid4()
-      message_server_instance_number = "00"
-      abap_host_id = UUID.uuid4()
-      abap_instance_number = "01"
-
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          sid: db_sid,
-          instance_number: db_instance_number_1,
-          system_replication: "Primary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_2,
-          system_replication: "Secondary",
-          sid: db_sid
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "MESSAGESERVER|ENQUE",
-          host_id: message_server_host_id,
-          instance_number: message_server_instance_number,
-          sid: application_sid
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "ABAP|GATEWAY|ICMAN|IGS",
-          host_id: abap_host_id,
-          instance_number: abap_instance_number,
-          sid: application_sid
-        ),
-        build(
-          :sap_system_registered_event,
-          sap_system_id: sap_system_id,
-          sid: application_sid
-        ),
-        build(:database_instance_deregistered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          instance_number: db_instance_number_1,
-          deregistered_at: deregistered_at
-        ),
-        build(:database_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        ),
-        build(:sap_system_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        )
-      ]
-
-      %{features: features, instance_number: instance_number, health: health} =
-        command =
-        build(:register_database_instance_command,
-          system_replication: "Primary",
-          sid: db_sid,
-          sap_system_id: sap_system_id
-        )
-
-      assert_events_and_state(
-        initial_events,
-        command,
-        [
-          %DatabaseInstanceRegistered{
-            sap_system_id: sap_system_id,
-            sid: db_sid,
-            tenant: command.tenant,
-            instance_number: command.instance_number,
-            instance_hostname: command.instance_hostname,
-            features: command.features,
-            http_port: command.http_port,
-            https_port: command.https_port,
-            start_priority: command.start_priority,
-            host_id: command.host_id,
-            system_replication: command.system_replication,
-            system_replication_status: command.system_replication_status,
-            health: command.health
-          },
-          %DatabaseRestored{
-            sap_system_id: sap_system_id,
-            health: command.health
-          }
-        ],
-        fn sap_system ->
-          assert Kernel.length(sap_system.database.instances) == 2
-
-          assert %SapSystem{
-                   deregistered_at: ^deregistered_at,
-                   database: %Database{
-                     deregistered_at: nil,
-                     sid: ^db_sid,
-                     instances: [
-                       %Instance{
-                         sid: ^db_sid,
-                         instance_number: ^instance_number,
-                         features: ^features,
-                         health: ^health,
-                         system_replication: "Primary"
-                       },
-                       %Instance{}
-                     ]
-                   }
-                 } = sap_system
-        end
-      )
-    end
-
     test "should not restore a sap system when no abap/messageserver instances are present" do
       sap_system_id = UUID.uuid4()
 
-      primary_database_host_id = UUID.uuid4()
-      secondary_database_host_id = UUID.uuid4()
+      database_host_id = UUID.uuid4()
 
       deregistered_at = DateTime.utc_now()
 
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
-
-      db_sid = fake_sid()
       application_sid = fake_sid()
 
       message_server_host_id = UUID.uuid4()
@@ -2742,27 +1272,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       abap_instance_number = "01"
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          sid: db_sid,
-          instance_number: db_instance_number_1,
-          system_replication: "Primary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_2,
-          system_replication: "Secondary",
-          sid: db_sid
-        ),
         build(
           :application_instance_registered_event,
           sap_system_id: sap_system_id,
@@ -2783,16 +1292,6 @@ defmodule Trento.SapSystems.SapSystemTest do
           :sap_system_registered_event,
           sap_system_id: sap_system_id,
           sid: application_sid
-        ),
-        build(:database_instance_deregistered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          instance_number: db_instance_number_1,
-          deregistered_at: deregistered_at
-        ),
-        build(:database_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
         ),
         build(:sap_system_deregistered_event,
           sap_system_id: sap_system_id,
@@ -2803,14 +1302,6 @@ defmodule Trento.SapSystems.SapSystemTest do
           deregistered_at: deregistered_at,
           instance_number: message_server_instance_number,
           host_id: message_server_host_id
-        ),
-        build(:database_instance_registered_event,
-          system_replication: "Primary",
-          sid: db_sid,
-          sap_system_id: sap_system_id
-        ),
-        build(:database_restored_event,
-          sap_system_id: sap_system_id
         )
       ]
 
@@ -2819,7 +1310,7 @@ defmodule Trento.SapSystems.SapSystemTest do
           :register_application_instance_command,
           sap_system_id: sap_system_id,
           sid: application_sid,
-          db_host: primary_database_host_id,
+          db_host: database_host_id,
           features: "IGS"
         )
 
@@ -2842,11 +1333,7 @@ defmodule Trento.SapSystems.SapSystemTest do
         ],
         fn sap_system ->
           assert %SapSystem{
-                   deregistered_at: ^deregistered_at,
-                   database: %Database{
-                     deregistered_at: nil,
-                     sid: ^db_sid
-                   }
+                   deregistered_at: ^deregistered_at
                  } = sap_system
         end
       )
@@ -2855,15 +1342,10 @@ defmodule Trento.SapSystems.SapSystemTest do
     test "should restore a sap system when abap/messageserver instances are present" do
       sap_system_id = UUID.uuid4()
 
-      primary_database_host_id = UUID.uuid4()
-      secondary_database_host_id = UUID.uuid4()
+      database_host_id = UUID.uuid4()
 
       deregistered_at = DateTime.utc_now()
 
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
-
-      db_sid = fake_sid()
       application_sid = fake_sid()
 
       message_server_host_id = UUID.uuid4()
@@ -2872,27 +1354,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       abap_instance_number = "01"
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          sid: db_sid,
-          instance_number: db_instance_number_1,
-          system_replication: "Primary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_2,
-          system_replication: "Secondary",
-          sid: db_sid
-        ),
         build(
           :application_instance_registered_event,
           sap_system_id: sap_system_id,
@@ -2914,16 +1375,6 @@ defmodule Trento.SapSystems.SapSystemTest do
           sap_system_id: sap_system_id,
           sid: application_sid
         ),
-        build(:database_instance_deregistered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          instance_number: db_instance_number_1,
-          deregistered_at: deregistered_at
-        ),
-        build(:database_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        ),
         build(:sap_system_deregistered_event,
           sap_system_id: sap_system_id,
           deregistered_at: deregistered_at
@@ -2933,14 +1384,6 @@ defmodule Trento.SapSystems.SapSystemTest do
           deregistered_at: deregistered_at,
           instance_number: message_server_instance_number,
           host_id: message_server_host_id
-        ),
-        build(:database_instance_registered_event,
-          system_replication: "Primary",
-          sid: db_sid,
-          sap_system_id: sap_system_id
-        ),
-        build(:database_restored_event,
-          sap_system_id: sap_system_id
         )
       ]
 
@@ -2949,7 +1392,7 @@ defmodule Trento.SapSystems.SapSystemTest do
           :register_application_instance_command,
           sap_system_id: sap_system_id,
           sid: application_sid,
-          db_host: primary_database_host_id,
+          db_host: database_host_id,
           features: "MESSAGESERVER"
         )
 
@@ -2978,133 +1421,7 @@ defmodule Trento.SapSystems.SapSystemTest do
         ],
         fn sap_system ->
           assert %SapSystem{
-                   deregistered_at: nil,
-                   database: %Database{
-                     deregistered_at: nil,
-                     sid: ^db_sid
-                   }
-                 } = sap_system
-        end
-      )
-    end
-
-    test "should not restore a sap system if the database is not present and messageserver/abap instance are present" do
-      sap_system_id = UUID.uuid4()
-
-      primary_database_host_id = UUID.uuid4()
-      secondary_database_host_id = UUID.uuid4()
-
-      deregistered_at = DateTime.utc_now()
-
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
-
-      db_sid = fake_sid()
-      application_sid = fake_sid()
-
-      message_server_host_id = UUID.uuid4()
-      message_server_instance_number = "00"
-      abap_host_id = UUID.uuid4()
-      abap_instance_number = "01"
-
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          sid: db_sid,
-          instance_number: db_instance_number_1,
-          system_replication: "Primary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_2,
-          system_replication: "Secondary",
-          sid: db_sid
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "MESSAGESERVER|ENQUE",
-          host_id: message_server_host_id,
-          instance_number: message_server_instance_number,
-          sid: application_sid
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "ABAP|GATEWAY|ICMAN|IGS",
-          host_id: abap_host_id,
-          instance_number: abap_instance_number,
-          sid: application_sid
-        ),
-        build(
-          :sap_system_registered_event,
-          sap_system_id: sap_system_id,
-          sid: application_sid
-        ),
-        build(:database_instance_deregistered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          instance_number: db_instance_number_1,
-          deregistered_at: deregistered_at
-        ),
-        build(:database_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        ),
-        build(:sap_system_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        ),
-        build(:application_instance_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at,
-          instance_number: message_server_instance_number,
-          host_id: message_server_host_id
-        )
-      ]
-
-      command =
-        build(
-          :register_application_instance_command,
-          sap_system_id: sap_system_id,
-          sid: application_sid,
-          db_host: primary_database_host_id,
-          features: "MESSAGESERVER"
-        )
-
-      assert_events_and_state(
-        initial_events,
-        command,
-        [
-          %ApplicationInstanceRegistered{
-            sap_system_id: sap_system_id,
-            sid: application_sid,
-            host_id: command.host_id,
-            instance_number: command.instance_number,
-            instance_hostname: command.instance_hostname,
-            features: command.features,
-            http_port: command.http_port,
-            https_port: command.https_port,
-            start_priority: command.start_priority,
-            health: command.health
-          }
-        ],
-        fn sap_system ->
-          assert %SapSystem{
-                   deregistered_at: ^deregistered_at,
-                   database: %Database{
-                     deregistered_at: ^deregistered_at,
-                     sid: ^db_sid
-                   }
+                   deregistered_at: nil
                  } = sap_system
         end
       )
@@ -3113,15 +1430,8 @@ defmodule Trento.SapSystems.SapSystemTest do
     test "should reject all the commands except for the registration/instance deregistration ones, when the SAP system is deregistered" do
       sap_system_id = UUID.uuid4()
 
-      primary_database_host_id = UUID.uuid4()
-      secondary_database_host_id = UUID.uuid4()
-
       deregistered_at = DateTime.utc_now()
 
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
-
-      db_sid = fake_sid()
       application_sid = fake_sid()
 
       message_server_host_id = UUID.uuid4()
@@ -3130,27 +1440,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       abap_instance_number = "01"
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          sid: db_sid,
-          instance_number: db_instance_number_1,
-          system_replication: "Primary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_2,
-          system_replication: "Secondary",
-          sid: db_sid
-        ),
         build(
           :application_instance_registered_event,
           sap_system_id: sap_system_id,
@@ -3172,16 +1461,6 @@ defmodule Trento.SapSystems.SapSystemTest do
           sap_system_id: sap_system_id,
           sid: application_sid
         ),
-        build(:database_instance_deregistered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          instance_number: db_instance_number_1,
-          deregistered_at: deregistered_at
-        ),
-        build(:database_deregistered_event,
-          sap_system_id: sap_system_id,
-          deregistered_at: deregistered_at
-        ),
         build(:sap_system_deregistered_event,
           sap_system_id: sap_system_id,
           deregistered_at: deregistered_at
@@ -3189,7 +1468,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       ]
 
       commands_to_accept = [
-        build(:register_database_instance_command),
         build(:register_application_instance_command),
         build(:rollup_sap_system_command)
       ]
@@ -3200,517 +1478,22 @@ defmodule Trento.SapSystems.SapSystemTest do
       end
     end
 
-    test "should deregister a Database and SAP system when the Primary database instance is removed" do
-      sap_system_id = UUID.uuid4()
-
-      primary_database_host_id = UUID.uuid4()
-      secondary_database_host_id = UUID.uuid4()
-
-      deregistered_at = DateTime.utc_now()
-
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
-
-      db_sid = fake_sid()
-      application_sid = fake_sid()
-
-      message_server_host_id = UUID.uuid4()
-      message_server_instance_number = "00"
-      abap_host_id = UUID.uuid4()
-      abap_instance_number = "01"
-
-      assert_events_and_state(
-        [
-          build(
-            :database_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            host_id: primary_database_host_id,
-            sid: db_sid,
-            instance_number: db_instance_number_1,
-            system_replication: "Primary"
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            host_id: secondary_database_host_id,
-            instance_number: db_instance_number_2,
-            system_replication: "Secondary",
-            sid: db_sid
-          ),
-          build(
-            :application_instance_registered_event,
-            sap_system_id: sap_system_id,
-            features: "MESSAGESERVER|ENQUE",
-            host_id: message_server_host_id,
-            instance_number: message_server_instance_number,
-            sid: application_sid
-          ),
-          build(
-            :application_instance_registered_event,
-            sap_system_id: sap_system_id,
-            features: "ABAP|GATEWAY|ICMAN|IGS",
-            host_id: abap_host_id,
-            instance_number: abap_instance_number,
-            sid: application_sid
-          ),
-          build(
-            :sap_system_registered_event,
-            sap_system_id: sap_system_id,
-            sid: application_sid
-          )
-        ],
-        %DeregisterDatabaseInstance{
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          instance_number: db_instance_number_1,
-          deregistered_at: deregistered_at
-        },
-        [
-          %DatabaseInstanceDeregistered{
-            sap_system_id: sap_system_id,
-            host_id: primary_database_host_id,
-            instance_number: db_instance_number_1,
-            deregistered_at: deregistered_at
-          },
-          %DatabaseDeregistered{
-            sap_system_id: sap_system_id,
-            deregistered_at: deregistered_at
-          },
-          %SapSystemDeregistered{
-            sap_system_id: sap_system_id,
-            deregistered_at: deregistered_at
-          }
-        ],
-        fn sap_system ->
-          assert %SapSystem{
-                   sid: ^application_sid,
-                   database: %Database{
-                     sid: ^db_sid,
-                     instances: [
-                       %Instance{
-                         instance_number: ^db_instance_number_2,
-                         sid: ^db_sid,
-                         host_id: ^secondary_database_host_id
-                       }
-                     ],
-                     health: :passing,
-                     deregistered_at: ^deregistered_at
-                   },
-                   application: %Application{
-                     sid: ^application_sid,
-                     instances: [
-                       %Instance{instance_number: ^abap_instance_number, sid: ^application_sid},
-                       %Instance{
-                         instance_number: ^message_server_instance_number,
-                         sid: ^application_sid
-                       }
-                     ]
-                   },
-                   deregistered_at: ^deregistered_at
-                 } = sap_system
-        end
-      )
-    end
-
-    test "should deregister a secondary DB instance, no SAP system registered." do
-      sap_system_id = UUID.uuid4()
-      deregistered_at = DateTime.utc_now()
-
-      instance_number_1 = "00"
-      instance_number_2 = "01"
-
-      db_sid = fake_sid()
-
-      primary_db_host_id = UUID.uuid4()
-      secondary_db_host_id = UUID.uuid4()
-
-      assert_events_and_state(
-        [
-          build(
-            :database_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            system_replication: "Primary",
-            sid: db_sid,
-            instance_number: instance_number_1,
-            host_id: primary_db_host_id
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            system_replication: "Secondary",
-            sid: db_sid,
-            host_id: secondary_db_host_id,
-            instance_number: instance_number_2
-          )
-        ],
-        %DeregisterDatabaseInstance{
-          sap_system_id: sap_system_id,
-          host_id: secondary_db_host_id,
-          instance_number: instance_number_2,
-          deregistered_at: deregistered_at
-        },
-        %DatabaseInstanceDeregistered{
-          sap_system_id: sap_system_id,
-          host_id: secondary_db_host_id,
-          instance_number: instance_number_2,
-          deregistered_at: deregistered_at
-        },
-        fn sap_system ->
-          assert %SapSystem{
-                   sid: nil,
-                   database: %Database{
-                     sid: ^db_sid,
-                     deregistered_at: nil,
-                     instances: [%Instance{instance_number: ^instance_number_1}]
-                   }
-                 } = sap_system
-        end
-      )
-    end
-
-    test "should deregister the only database instance and deregister the entire database, no SAP system registered, system replication disabled" do
-      sap_system_id = UUID.uuid4()
-      database_host_id = UUID.uuid4()
-      deregistered_at = DateTime.utc_now()
-      db_instance_number_1 = "00"
-      db_sid = fake_sid()
-
-      assert_events_and_state(
-        [
-          build(
-            :database_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            host_id: database_host_id,
-            instance_number: db_instance_number_1,
-            system_replication: nil,
-            sid: db_sid
-          )
-        ],
-        %DeregisterDatabaseInstance{
-          sap_system_id: sap_system_id,
-          host_id: database_host_id,
-          instance_number: db_instance_number_1,
-          deregistered_at: deregistered_at
-        },
-        [
-          %DatabaseInstanceDeregistered{
-            sap_system_id: sap_system_id,
-            host_id: database_host_id,
-            instance_number: db_instance_number_1,
-            deregistered_at: deregistered_at
-          },
-          %DatabaseDeregistered{
-            sap_system_id: sap_system_id,
-            deregistered_at: deregistered_at
-          }
-        ],
-        fn sap_system ->
-          assert %SapSystem{
-                   sid: nil,
-                   database: %Database{
-                     sid: ^db_sid,
-                     instances: [],
-                     health: :passing,
-                     deregistered_at: ^deregistered_at
-                   },
-                   application: nil,
-                   deregistered_at: nil
-                 } = sap_system
-        end
-      )
-    end
-
-    test "should deregister a single DB instance of two if no SR enabled, SAP system not registered" do
-      sap_system_id = UUID.uuid4()
-      database_host_id = UUID.uuid4()
-      second_database_host_id = UUID.uuid4()
-      deregistered_at = DateTime.utc_now()
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
-
-      db_sid = fake_sid()
-
-      assert_events_and_state(
-        [
-          build(
-            :database_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            host_id: database_host_id,
-            instance_number: db_instance_number_1,
-            system_replication: nil,
-            sid: db_sid
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            host_id: second_database_host_id,
-            instance_number: db_instance_number_2,
-            system_replication: nil,
-            sid: db_sid
-          )
-        ],
-        %DeregisterDatabaseInstance{
-          sap_system_id: sap_system_id,
-          host_id: database_host_id,
-          instance_number: db_instance_number_1,
-          deregistered_at: deregistered_at
-        },
-        %DatabaseInstanceDeregistered{
-          sap_system_id: sap_system_id,
-          host_id: database_host_id,
-          instance_number: db_instance_number_1,
-          deregistered_at: deregistered_at
-        },
-        fn sap_system ->
-          assert %SapSystem{
-                   database: %Database{
-                     sid: ^db_sid,
-                     deregistered_at: nil,
-                     instances: [%Instance{instance_number: ^db_instance_number_2}]
-                   }
-                 } = sap_system
-        end
-      )
-    end
-
-    test "should deregister the primary instance, the entire database and then the secondary instance, SAP system not registered" do
-      sap_system_id = UUID.uuid4()
-      primary_database_host_id = UUID.uuid4()
-      secondary_database_host_id = UUID.uuid4()
-      deregistered_at = DateTime.utc_now()
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
-
-      db_sid = fake_sid()
-
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          sid: db_sid,
-          instance_number: db_instance_number_1,
-          system_replication: "Primary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_2,
-          system_replication: "Secondary"
-        )
-      ]
-
-      assert_events(
-        initial_events,
-        [
-          %DeregisterDatabaseInstance{
-            sap_system_id: sap_system_id,
-            host_id: primary_database_host_id,
-            instance_number: db_instance_number_1,
-            deregistered_at: deregistered_at
-          },
-          %DeregisterDatabaseInstance{
-            sap_system_id: sap_system_id,
-            host_id: secondary_database_host_id,
-            instance_number: db_instance_number_2,
-            deregistered_at: deregistered_at
-          }
-        ],
-        [
-          %DatabaseInstanceDeregistered{
-            sap_system_id: sap_system_id,
-            host_id: primary_database_host_id,
-            instance_number: db_instance_number_1,
-            deregistered_at: deregistered_at
-          },
-          %DatabaseDeregistered{
-            sap_system_id: sap_system_id,
-            deregistered_at: deregistered_at
-          },
-          %DatabaseInstanceDeregistered{
-            sap_system_id: sap_system_id,
-            host_id: secondary_database_host_id,
-            instance_number: db_instance_number_2,
-            deregistered_at: deregistered_at
-          }
-        ]
-      )
-    end
-
-    test "should correctly deregister the database in a scale out scenario, with two primary and two secondary, no SAP system registered" do
-      sap_system_id = UUID.uuid4()
-      first_primary_database_host_id = UUID.uuid4()
-      other_primary_database_host_id = UUID.uuid4()
-
-      secondary_database_host_id = UUID.uuid4()
-      other_secondary_database_host_id = UUID.uuid4()
-
-      deregistered_at = DateTime.utc_now()
-
-      db_instance_number_1 = "00"
-      db_instance_number_2 = "01"
-      db_instance_number_3 = "02"
-      db_instance_number_4 = "03"
-
-      db_sid = fake_sid()
-
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: first_primary_database_host_id,
-          sid: db_sid,
-          instance_number: db_instance_number_1,
-          system_replication: "Primary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: other_primary_database_host_id,
-          sid: db_sid,
-          instance_number: db_instance_number_2,
-          system_replication: "Primary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: db_instance_number_3,
-          system_replication: "Secondary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: other_secondary_database_host_id,
-          instance_number: db_instance_number_4,
-          system_replication: "Secondary"
-        )
-      ]
-
-      assert_events(
-        initial_events,
-        [
-          %DeregisterDatabaseInstance{
-            sap_system_id: sap_system_id,
-            host_id: first_primary_database_host_id,
-            instance_number: db_instance_number_1,
-            deregistered_at: deregistered_at
-          },
-          %DeregisterDatabaseInstance{
-            sap_system_id: sap_system_id,
-            host_id: other_primary_database_host_id,
-            instance_number: db_instance_number_2,
-            deregistered_at: deregistered_at
-          },
-          %DeregisterDatabaseInstance{
-            sap_system_id: sap_system_id,
-            host_id: secondary_database_host_id,
-            instance_number: db_instance_number_3,
-            deregistered_at: deregistered_at
-          },
-          %DeregisterDatabaseInstance{
-            sap_system_id: sap_system_id,
-            host_id: other_secondary_database_host_id,
-            instance_number: db_instance_number_4,
-            deregistered_at: deregistered_at
-          }
-        ],
-        [
-          %DatabaseInstanceDeregistered{
-            sap_system_id: sap_system_id,
-            host_id: first_primary_database_host_id,
-            instance_number: db_instance_number_1,
-            deregistered_at: deregistered_at
-          },
-          %DatabaseInstanceDeregistered{
-            sap_system_id: sap_system_id,
-            host_id: other_primary_database_host_id,
-            instance_number: db_instance_number_2,
-            deregistered_at: deregistered_at
-          },
-          %DatabaseDeregistered{
-            sap_system_id: sap_system_id,
-            deregistered_at: deregistered_at
-          },
-          %DatabaseInstanceDeregistered{
-            sap_system_id: sap_system_id,
-            host_id: secondary_database_host_id,
-            instance_number: db_instance_number_3,
-            deregistered_at: deregistered_at
-          },
-          %DatabaseInstanceDeregistered{
-            sap_system_id: sap_system_id,
-            host_id: other_secondary_database_host_id,
-            instance_number: db_instance_number_4,
-            deregistered_at: deregistered_at
-          }
-        ]
-      )
-    end
-
     test "should deregister an ENQREP Application Instance, SAP system registered" do
       sap_system_id = UUID.uuid4()
       deregistered_at = DateTime.utc_now()
 
-      database_host_id = UUID.uuid4()
       message_server_host_id = UUID.uuid4()
       abap_host_id = UUID.uuid4()
       enqrep_host_id = UUID.uuid4()
 
-      database_instance_number = "00"
       message_server_instance_number = "01"
       abap_instance_number = "02"
       enqrep_server_instance_number = "03"
 
-      db_sid = fake_sid()
       application_sid = fake_sid()
 
       assert_events_and_state(
         [
-          build(
-            :database_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            host_id: database_host_id,
-            sid: db_sid,
-            instance_number: database_instance_number,
-            system_replication: nil
-          ),
           build(
             :application_instance_registered_event,
             sap_system_id: sap_system_id,
@@ -3759,29 +1542,16 @@ defmodule Trento.SapSystems.SapSystemTest do
           assert %SapSystem{
                    sid: ^application_sid,
                    deregistered_at: nil,
-                   database: %Database{
-                     sid: ^db_sid,
-                     deregistered_at: nil,
-                     instances: [
-                       %Instance{
-                         instance_number: ^database_instance_number,
-                         host_id: ^database_host_id
-                       }
-                     ]
-                   },
-                   application: %Application{
-                     sid: ^application_sid,
-                     instances: [
-                       %Instance{
-                         host_id: ^abap_host_id,
-                         instance_number: ^abap_instance_number
-                       },
-                       %Instance{
-                         host_id: ^message_server_host_id,
-                         instance_number: ^message_server_instance_number
-                       }
-                     ]
-                   }
+                   instances: [
+                     %Instance{
+                       host_id: ^abap_host_id,
+                       instance_number: ^abap_instance_number
+                     },
+                     %Instance{
+                       host_id: ^message_server_host_id,
+                       instance_number: ^message_server_instance_number
+                     }
+                   ]
                  } = sap_system
         end
       )
@@ -3792,15 +1562,12 @@ defmodule Trento.SapSystems.SapSystemTest do
       deregistered_at = DateTime.utc_now()
 
       application_sid = fake_sid()
-      db_sid = fake_sid()
 
-      database_host_id = UUID.uuid4()
       message_server_host_id = UUID.uuid4()
       abap_host_id = UUID.uuid4()
       abap_2_host_id = UUID.uuid4()
       enqrep_host_id = UUID.uuid4()
 
-      database_instance_number = "00"
       message_server_instance_number = "01"
       abap_instance_number = "02"
       abap_2_instance_number = "03"
@@ -3808,19 +1575,6 @@ defmodule Trento.SapSystems.SapSystemTest do
 
       assert_events_and_state(
         [
-          build(
-            :database_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            host_id: database_host_id,
-            sid: db_sid,
-            instance_number: database_instance_number,
-            system_replication: nil
-          ),
           build(
             :application_instance_registered_event,
             sap_system_id: sap_system_id,
@@ -3877,33 +1631,20 @@ defmodule Trento.SapSystems.SapSystemTest do
           assert %SapSystem{
                    sid: ^application_sid,
                    deregistered_at: nil,
-                   database: %Database{
-                     sid: ^db_sid,
-                     deregistered_at: nil,
-                     instances: [
-                       %Instance{
-                         instance_number: ^database_instance_number,
-                         host_id: ^database_host_id
-                       }
-                     ]
-                   },
-                   application: %Application{
-                     sid: ^application_sid,
-                     instances: [
-                       %Instance{
-                         host_id: ^enqrep_host_id,
-                         instance_number: ^enqrep_server_instance_number
-                       },
-                       %Instance{
-                         host_id: ^abap_host_id,
-                         instance_number: ^abap_instance_number
-                       },
-                       %Instance{
-                         host_id: ^message_server_host_id,
-                         instance_number: ^message_server_instance_number
-                       }
-                     ]
-                   }
+                   instances: [
+                     %Instance{
+                       host_id: ^enqrep_host_id,
+                       instance_number: ^enqrep_server_instance_number
+                     },
+                     %Instance{
+                       host_id: ^abap_host_id,
+                       instance_number: ^abap_instance_number
+                     },
+                     %Instance{
+                       host_id: ^message_server_host_id,
+                       instance_number: ^message_server_instance_number
+                     }
+                   ]
                  } = sap_system
         end
       )
@@ -3912,33 +1653,18 @@ defmodule Trento.SapSystems.SapSystemTest do
     test "should deregister last ABAP Application Instance and deregister SAP System" do
       sap_system_id = UUID.uuid4()
       deregistered_at = DateTime.utc_now()
-      db_sid = fake_sid()
       application_sid = fake_sid()
 
-      database_host_id = UUID.uuid4()
       message_server_host_id = UUID.uuid4()
       abap_host_id = UUID.uuid4()
       enqrep_host_id = UUID.uuid4()
 
-      database_instance_number = "00"
       message_server_instance_number = "01"
       abap_instance_number = "02"
       enqrep_server_instance_number = "03"
 
       assert_events_and_state(
         [
-          build(
-            :database_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            host_id: database_host_id,
-            sid: db_sid,
-            instance_number: database_instance_number
-          ),
           build(
             :application_instance_registered_event,
             sap_system_id: sap_system_id,
@@ -3990,28 +1716,16 @@ defmodule Trento.SapSystems.SapSystemTest do
         fn sap_system ->
           assert %SapSystem{
                    sid: ^application_sid,
-                   database: %Database{
-                     deregistered_at: nil,
-                     sid: ^db_sid,
-                     instances: [
-                       %Instance{
-                         instance_number: ^database_instance_number,
-                         host_id: ^database_host_id
-                       }
-                     ]
-                   },
-                   application: %Application{
-                     instances: [
-                       %Instance{
-                         instance_number: ^enqrep_server_instance_number,
-                         host_id: ^enqrep_host_id
-                       },
-                       %Instance{
-                         instance_number: ^message_server_instance_number,
-                         host_id: ^message_server_host_id
-                       }
-                     ]
-                   },
+                   instances: [
+                     %Instance{
+                       instance_number: ^enqrep_server_instance_number,
+                       host_id: ^enqrep_host_id
+                     },
+                     %Instance{
+                       instance_number: ^message_server_instance_number,
+                       host_id: ^message_server_host_id
+                     }
+                   ],
                    deregistered_at: ^deregistered_at
                  } = sap_system
         end
@@ -4022,29 +1736,14 @@ defmodule Trento.SapSystems.SapSystemTest do
       sap_system_id = UUID.uuid4()
       deregistered_at = DateTime.utc_now()
 
-      database_host_id = UUID.uuid4()
       message_server_host_id = UUID.uuid4()
 
-      database_instance_number = "00"
       message_server_instance_number = "01"
 
       application_sid = fake_sid()
-      database_sid = fake_sid()
 
       assert_events_and_state(
         [
-          build(
-            :database_registered_event,
-            sid: database_sid,
-            sap_system_id: sap_system_id
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            sid: database_sid,
-            host_id: database_host_id,
-            instance_number: database_instance_number
-          ),
           build(
             :application_instance_registered_event,
             sap_system_id: sap_system_id,
@@ -4060,28 +1759,21 @@ defmodule Trento.SapSystems.SapSystemTest do
           instance_number: message_server_instance_number,
           deregistered_at: deregistered_at
         },
-        %ApplicationInstanceDeregistered{
-          sap_system_id: sap_system_id,
-          host_id: message_server_host_id,
-          instance_number: message_server_instance_number,
-          deregistered_at: deregistered_at
-        },
+        [
+          %ApplicationInstanceDeregistered{
+            sap_system_id: sap_system_id,
+            host_id: message_server_host_id,
+            instance_number: message_server_instance_number,
+            deregistered_at: deregistered_at
+          },
+          %SapSystemTombstoned{
+            sap_system_id: sap_system_id
+          }
+        ],
         fn sap_system ->
           assert %SapSystem{
                    sid: nil,
-                   database: %Database{
-                     sid: ^database_sid,
-                     deregistered_at: nil,
-                     instances: [
-                       %Instance{
-                         instance_number: ^database_instance_number,
-                         host_id: ^database_host_id
-                       }
-                     ]
-                   },
-                   application: %Application{
-                     instances: []
-                   },
+                   instances: [],
                    deregistered_at: nil
                  } = sap_system
         end
@@ -4092,33 +1784,18 @@ defmodule Trento.SapSystems.SapSystemTest do
       sap_system_id = UUID.uuid4()
       deregistered_at = DateTime.utc_now()
 
-      database_host_id = UUID.uuid4()
       message_server_host_id = UUID.uuid4()
       abap_host_id = UUID.uuid4()
       enqrep_host_id = UUID.uuid4()
 
-      database_instance_number = "00"
       message_server_instance_number = "01"
       abap_instance_number = "02"
       enqrep_server_instance_number = "03"
 
-      db_sid = fake_sid()
       application_sid = fake_sid()
 
       assert_events_and_state(
         [
-          build(
-            :database_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            host_id: database_host_id,
-            instance_number: database_instance_number,
-            sid: db_sid
-          ),
           build(
             :application_instance_registered_event,
             sap_system_id: sap_system_id,
@@ -4169,27 +1846,16 @@ defmodule Trento.SapSystems.SapSystemTest do
         ],
         fn sap_system ->
           assert %SapSystem{
-                   database: %Database{
-                     sid: ^db_sid,
-                     instances: [
-                       %Instance{
-                         instance_number: ^database_instance_number,
-                         host_id: ^database_host_id
-                       }
-                     ]
-                   },
-                   application: %Application{
-                     instances: [
-                       %Instance{
-                         host_id: ^enqrep_host_id,
-                         instance_number: ^enqrep_server_instance_number
-                       },
-                       %Instance{
-                         host_id: ^abap_host_id,
-                         instance_number: ^abap_instance_number
-                       }
-                     ]
-                   },
+                   instances: [
+                     %Instance{
+                       host_id: ^enqrep_host_id,
+                       instance_number: ^enqrep_server_instance_number
+                     },
+                     %Instance{
+                       host_id: ^abap_host_id,
+                       instance_number: ^abap_instance_number
+                     }
+                   ],
                    deregistered_at: ^deregistered_at,
                    sid: ^application_sid
                  } = sap_system
@@ -4197,130 +1863,11 @@ defmodule Trento.SapSystems.SapSystemTest do
       )
     end
 
-    test "should deregister the primary instance of database, the SAP system and deregister the ABAP application instance" do
-      sap_system_id = UUID.uuid4()
-      deregistered_at = DateTime.utc_now()
-
-      primary_database_host_id = UUID.uuid4()
-      secondary_database_host_id = UUID.uuid4()
-
-      message_server_host_id = UUID.uuid4()
-      abap_host_id = UUID.uuid4()
-
-      database_instance_number_1 = "00"
-      database_instance_number_2 = "00"
-
-      message_server_instance_number = "01"
-      abap_instance_number = "02"
-
-      db_sid = fake_sid()
-      application_sid = fake_sid()
-
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: db_sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: primary_database_host_id,
-          sid: db_sid,
-          instance_number: database_instance_number_1,
-          system_replication: "Primary"
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          host_id: secondary_database_host_id,
-          instance_number: database_instance_number_2,
-          system_replication: "Secondary"
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "MESSAGESERVER|ENQUE",
-          sid: application_sid,
-          host_id: message_server_host_id,
-          instance_number: message_server_instance_number
-        ),
-        build(
-          :application_instance_registered_event,
-          sap_system_id: sap_system_id,
-          features: "ABAP|GATEWAY|ICMAN|IGS",
-          sid: application_sid,
-          host_id: abap_host_id,
-          instance_number: abap_instance_number
-        ),
-        build(
-          :sap_system_registered_event,
-          sap_system_id: sap_system_id,
-          sid: application_sid
-        )
-      ]
-
-      assert_events(
-        initial_events,
-        [
-          %DeregisterDatabaseInstance{
-            sap_system_id: sap_system_id,
-            host_id: primary_database_host_id,
-            instance_number: database_instance_number_1,
-            deregistered_at: deregistered_at
-          },
-          %DeregisterApplicationInstance{
-            sap_system_id: sap_system_id,
-            host_id: message_server_host_id,
-            instance_number: message_server_instance_number,
-            deregistered_at: deregistered_at
-          }
-        ],
-        [
-          %DatabaseInstanceDeregistered{
-            sap_system_id: sap_system_id,
-            host_id: primary_database_host_id,
-            instance_number: database_instance_number_1,
-            deregistered_at: deregistered_at
-          },
-          %DatabaseDeregistered{
-            sap_system_id: sap_system_id,
-            deregistered_at: deregistered_at
-          },
-          %SapSystemDeregistered{
-            sap_system_id: sap_system_id,
-            deregistered_at: deregistered_at
-          },
-          %ApplicationInstanceDeregistered{
-            sap_system_id: sap_system_id,
-            host_id: message_server_host_id,
-            instance_number: message_server_instance_number,
-            deregistered_at: deregistered_at
-          }
-        ]
-      )
-    end
-
     test "should not deregister a not registered application instance" do
       sap_system_id = UUID.uuid4()
-      db_sid = fake_sid()
 
       assert_error(
-        [
-          build(
-            :database_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid,
-            host_id: UUID.uuid4(),
-            instance_number: "00",
-            system_replication: "Primary"
-          )
-        ],
+        [],
         [
           %DeregisterApplicationInstance{
             sap_system_id: sap_system_id,
@@ -4329,32 +1876,18 @@ defmodule Trento.SapSystems.SapSystemTest do
             deregistered_at: DateTime.utc_now()
           }
         ],
-        {:error, :application_instance_not_registered}
+        {:error, :sap_system_not_registered}
       )
     end
 
     test "should not deregister an already deregistered application instance" do
       sap_system_id = UUID.uuid4()
-      db_sid = fake_sid()
       application_sid = fake_sid()
       deregistered_host_id = UUID.uuid4()
       deregistered_instance_number = "02"
 
       assert_error(
         [
-          build(
-            :database_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid,
-            host_id: UUID.uuid4(),
-            instance_number: "00",
-            system_replication: "Primary"
-          ),
           build(
             :application_instance_registered_event,
             sap_system_id: sap_system_id,
@@ -4395,169 +1928,9 @@ defmodule Trento.SapSystems.SapSystemTest do
         {:error, :application_instance_not_registered}
       )
     end
-
-    test "should not deregister a not registered database instance" do
-      sap_system_id = UUID.uuid4()
-      db_sid = fake_sid()
-
-      assert_error(
-        [
-          build(
-            :database_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid,
-            host_id: UUID.uuid4(),
-            instance_number: "00",
-            system_replication: "Primary"
-          )
-        ],
-        [
-          %DeregisterDatabaseInstance{
-            sap_system_id: sap_system_id,
-            host_id: UUID.uuid4(),
-            instance_number: "01",
-            deregistered_at: DateTime.utc_now()
-          }
-        ],
-        {:error, :database_instance_not_registered}
-      )
-    end
-
-    test "should not deregister an already deregistered database instance" do
-      sap_system_id = UUID.uuid4()
-      db_sid = fake_sid()
-      deregistered_host_id = UUID.uuid4()
-      deregistered_instance_number = "01"
-
-      assert_error(
-        [
-          build(
-            :database_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid
-          ),
-          build(
-            :database_instance_registered_event,
-            sap_system_id: sap_system_id,
-            sid: db_sid,
-            host_id: deregistered_host_id,
-            instance_number: deregistered_instance_number,
-            system_replication: "Primary"
-          ),
-          build(
-            :database_instance_deregistered_event,
-            sap_system_id: sap_system_id,
-            host_id: deregistered_host_id,
-            instance_number: deregistered_instance_number,
-            deregistered_at: DateTime.utc_now()
-          )
-        ],
-        [
-          %DeregisterDatabaseInstance{
-            sap_system_id: sap_system_id,
-            host_id: deregistered_host_id,
-            instance_number: deregistered_instance_number,
-            deregistered_at: DateTime.utc_now()
-          }
-        ],
-        {:error, :database_instance_not_registered}
-      )
-    end
   end
 
   describe "instance marked absent/present" do
-    test "should mark as absent a previously registered database instance" do
-      sap_system_id = Faker.UUID.v4()
-      sid = fake_sid()
-      host_id = Faker.UUID.v4()
-      absent_db_instance_number = "01"
-      present_db_instance_number = "02"
-      absent_db_absent_at = DateTime.utc_now()
-
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          host_id: host_id,
-          instance_number: absent_db_instance_number,
-          system_replication: nil,
-          system_replication_status: nil
-        ),
-        build(
-          :database_instance_marked_absent_event,
-          sap_system_id: sap_system_id,
-          host_id: host_id,
-          instance_number: absent_db_instance_number,
-          absent_at: absent_db_absent_at
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          host_id: host_id,
-          instance_number: present_db_instance_number,
-          system_replication: nil,
-          system_replication_status: nil
-        )
-      ]
-
-      absent_at = DateTime.utc_now()
-
-      assert_events_and_state(
-        initial_events,
-        [
-          %MarkDatabaseInstanceAbsent{
-            instance_number: absent_db_instance_number,
-            host_id: host_id,
-            sap_system_id: sap_system_id,
-            absent_at: absent_at
-          },
-          %MarkDatabaseInstanceAbsent{
-            instance_number: present_db_instance_number,
-            host_id: host_id,
-            sap_system_id: sap_system_id,
-            absent_at: absent_at
-          }
-        ],
-        [
-          %DatabaseInstanceMarkedAbsent{
-            instance_number: present_db_instance_number,
-            host_id: host_id,
-            sap_system_id: sap_system_id,
-            absent_at: absent_at
-          }
-        ],
-        fn state ->
-          assert %SapSystem{
-                   database: %Database{
-                     sid: ^sid,
-                     instances: [
-                       %Instance{
-                         instance_number: ^present_db_instance_number,
-                         absent_at: ^absent_at
-                       },
-                       %Instance{
-                         instance_number: ^absent_db_instance_number,
-                         absent_at: ^absent_db_absent_at
-                       }
-                     ]
-                   }
-                 } = state
-        end
-      )
-    end
-
     test "should mark as absent a previously registered application instance" do
       sap_system_id = Faker.UUID.v4()
       sid = fake_sid()
@@ -4567,26 +1940,17 @@ defmodule Trento.SapSystems.SapSystemTest do
       absent_app_absent_at = DateTime.utc_now()
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          host_id: host_id,
-          instance_number: "01",
-          system_replication: nil,
-          system_replication_status: nil
-        ),
         build(:application_instance_registered_event,
           sap_system_id: sap_system_id,
           sid: sid,
           host_id: host_id,
           instance_number: absent_message_server_instance_number,
           features: "MESSAGESERVER"
+        ),
+        build(
+          :sap_system_registered_event,
+          sap_system_id: sap_system_id,
+          sid: sid
         ),
         build(
           :application_instance_marked_absent_event,
@@ -4632,100 +1996,17 @@ defmodule Trento.SapSystems.SapSystemTest do
         ],
         fn state ->
           assert %SapSystem{
-                   application: %Application{
-                     sid: ^sid,
-                     instances: [
-                       %Instance{
-                         instance_number: ^present_app_instance_number,
-                         absent_at: ^absent_at
-                       },
-                       %Instance{
-                         instance_number: ^absent_message_server_instance_number,
-                         absent_at: ^absent_app_absent_at
-                       }
-                     ]
-                   }
-                 } = state
-        end
-      )
-    end
-
-    test "should mark as present an already registered, absent database instance" do
-      sap_system_id = Faker.UUID.v4()
-      sid = fake_sid()
-      host_id = Faker.UUID.v4()
-      absent_db_instance_number = "01"
-      present_db_instance_number = "02"
-
-      initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          host_id: host_id,
-          instance_number: absent_db_instance_number,
-          system_replication: nil,
-          system_replication_status: nil
-        ),
-        build(
-          :database_instance_marked_absent_event,
-          sap_system_id: sap_system_id,
-          host_id: host_id,
-          instance_number: absent_db_instance_number,
-          absent_at: DateTime.utc_now()
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          host_id: host_id,
-          instance_number: present_db_instance_number,
-          system_replication: nil,
-          system_replication_status: nil
-        )
-      ]
-
-      assert_events_and_state(
-        initial_events,
-        [
-          %RegisterDatabaseInstance{
-            sap_system_id: sap_system_id,
-            host_id: host_id,
-            instance_number: absent_db_instance_number,
-            health: :passing
-          },
-          %RegisterDatabaseInstance{
-            sap_system_id: sap_system_id,
-            host_id: host_id,
-            instance_number: present_db_instance_number,
-            health: :passing
-          }
-        ],
-        [
-          %DatabaseInstanceMarkedPresent{
-            instance_number: absent_db_instance_number,
-            host_id: host_id,
-            sap_system_id: sap_system_id
-          }
-        ],
-        fn state ->
-          assert %SapSystem{
-                   database: %Database{
-                     sid: ^sid,
-                     instances: [
-                       %Instance{
-                         absent_at: nil
-                       },
-                       %Instance{
-                         absent_at: nil
-                       }
-                     ]
-                   }
+                   sid: ^sid,
+                   instances: [
+                     %Instance{
+                       instance_number: ^present_app_instance_number,
+                       absent_at: ^absent_at
+                     },
+                     %Instance{
+                       instance_number: ^absent_message_server_instance_number,
+                       absent_at: ^absent_app_absent_at
+                     }
+                   ]
                  } = state
         end
       )
@@ -4740,20 +2021,6 @@ defmodule Trento.SapSystems.SapSystemTest do
       present_message_server_instance_number = "03"
 
       initial_events = [
-        build(
-          :database_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid
-        ),
-        build(
-          :database_instance_registered_event,
-          sap_system_id: sap_system_id,
-          sid: sid,
-          host_id: host_id,
-          instance_number: "01",
-          system_replication: nil,
-          system_replication_status: nil
-        ),
         build(:application_instance_registered_event,
           sap_system_id: sap_system_id,
           sid: sid,
@@ -4813,17 +2080,14 @@ defmodule Trento.SapSystems.SapSystemTest do
           assert %SapSystem{
                    sid: ^sid,
                    ensa_version: ^ensa_version,
-                   application: %Application{
-                     sid: ^sid,
-                     instances: [
-                       %Instance{
-                         absent_at: nil
-                       },
-                       %Instance{
-                         absent_at: nil
-                       }
-                     ]
-                   }
+                   instances: [
+                     %Instance{
+                       absent_at: nil
+                     },
+                     %Instance{
+                       absent_at: nil
+                     }
+                   ]
                  } = state
         end
       )
